+++ offchain/dispatcher/src/machine/mod.rs
@@ -10,7 +10,7 @@ use async_trait::async_trait;
 
 use self::rollups_broker::BrokerFacadeError;
 
-#[derive(Debug, Default)]
+#[derive(Debug, Clone, Copy, Default)]
 pub struct RollupStatus {
     pub inputs_sent_count: u64,
     pub last_event_is_finish_epoch: bool,